requirements.txt
tensorflow
pickle
flask
numpy==1.19.2
gunicorn
pillow